HADOOP-13786 Patch 028 * fixed up the javdoc/checkstyle complaints * metrics added, logged in MR test * improved cleanup of dirs, tests for this * use transfer manager for uploading the success data. I haven't done this for the actual block output stream, but I can, either in this patch or later * everything tested with s3guard & localb, as well as without it
feature wise, things are stablising. I'm taking this opportunity to do the big refactoring work to committers &c
I'm not doing any retries on operations. I think we must do that on the final commit of the put, the way S3ABlockOutputStream does. Why? A transient failure there loses the whole commit, hence job.

test-wise: reaching limit of what we can do without fault injection; next step is downstream stuff. My spark cloudexamples can do this, indeed, it will read in the success data and verify the committer used is the one invoked, so verifying that the committer is being used.