MAPREDUCE-7341. Intermediate Manifest Committer
This adds a new committer whose task commit operation is
a treewalk of the task attempt dir and the saving of a json
manifest summarising that tree in the job attempt dir.

job commit is a parallelised load of these files,
combined and parallelised mkdirs() of dest dirs,
then parallel renames to the destination.

Tests for file:// are in the mapreduce module, abfs adds
ITests, including terasorting.

It's targeted at spark on object stores without atomic
dir rename and whose dir delete may be O(descendants);
aspects of the design are suboptimal on HDFS.
And there's currently no attempt at job restart, which
the current MR committers support.