HADOOP-19388. Hadoop test of iceberg bulk delete
The new test is ITestIcebergBulkDelete; it is parameterized
Iceberg bulk delete enabled/disabled and s3a multipart
delete enabled/disabled.

There is a superclass contract test
  org.apache.fs.test.formats.AbstractIcebergDeleteTest
To support future stores which implement bulk delete.
This is currently a minimal superclass; all tests
are currently in ITestIcebergBulkDelete

HADOOP-19388. Building and testing with Iceberg 1.11.0-SNAPSHOT

Excluding the roaringbitmap dependency as it needs a new repo
and I don't trust it.